refactor conflictcontroller and tests
